parking-management/requirement.txt
Django==3.1.3
djangorestframework==3.12.1
Markdown==3.3.3
django-filter==2.4.0
